requirements.txt
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-Cors==3.0.10
flask-redis==0.4.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
redis==4.0.2
six==1.16.0
Werkzeug==2.0.2
wrapt==1.13.3
